Adding getResponse and deviceAction
Changing ListSensors.py name to tellstickHandler.py, getResponse method is to be used to get information about devices and sensors, to make it more modular added a f string appending the method. deviceAction method is to be used to perform actions on devices such as turn on and off device. There is a possibility to merge these functions into one because it almost replicated code except in params but might increase complexity.
Also added a debugging variable for printing data