Add feature getDevices and getSensors to tellstickHandler.py
getDevices method is to be used to get information of available device and display them in the android app, and also use that information to send action API calls to. getSensors for now will get information about sensors temperature. This information will be stored in a dictionary to map thing to data