Add dataHandler for retrieving and storing gesture data
Adding getGesture - retrieve data in form of a json object containing boolean array, the name of the gesture (set by the user), action - for example get temperature, device is the actuator/sensors you want to perform the action on.

Save gesture - saves the data in form of a json object in data.json in data folder. Storing as a number "1" gester which contains boolarray, name, action, device data