Updating EditGesture.java and Adding EditGestureToThing.java
Adding functionallity to select gesture to EditGesture.java file, then to complete the interaction added EditGestureToThing.java.

EditGestureToThing is very similar to AddGestureToThing.java, but instead it should display what values it was saved as so the user can change it. Added a button called change gesture to only change the gesture as the user might not want to change the gesture and only for example the action.