Create RunSSH for handling SSH connection
Using Ganymed SSH-2 to handle SSH connection to the raspberry pi. This will be used to read terminal output on the raspberry pi to then display data in the app. It has a run command, planning on adding arguments into the python code to provide information back to the raspberry pi