Add editGesture in dataHandler.py
Adding ability to edit saved gesture in data.json and then saves the result. There could be potential conflict with using name of the gesture, as the user could potentially have the same name for a gesture. So maybe a better way to deal with this is using deviceId, as it is unique.