Update AddGesture.java to display available sensors, and devices
Adding AsyncTask to send a SSH request to the raspberry pi and retrieve data from the available devices, and sensors. Currently not set with any command as the path in raspberry pi might change in future.

deviceList, and sensorList are used to append the values retrieved in the returnValueDevices/Sensors as the devices variable only passes as a reference.