Update AddGesture.java to read data from Raspberry PI
Removing the String[] sensors variable as it is no longer used.

Instead sending data from raspberry pi in form of a JSON object which can be easily read by Java's JSONObject. Then we Iterate over the keys to get the name of the device and add it to the ArrayList<String> sensorsList to be used in the ArrayAdapter.

Changing the openGestureActivity to putExtra in form of string of the data retrieved in the raspberry pi. Need to somehow distinguish if we send sensor data or device data as they are different.