Update gestureTracker.py to handle command line arguments
Adding two new functions to gestureTracker.py, one setGestureBoolArray that takes a gesture bool array and writes to the file gesture_bool_array.txt in data folder.

This function is linked to the getGestureBoolArray function that reads from the gesture_bool_array.txt in the data folder and returns the last known boolarray.

This is then integrated into the main.py file to have a --start command line argument to start capturing video feed. To be able to get the gesture bool array it needs to be saved in the gesture_bool_array.txt