Update AddGestureToThing.java to handle user confirm interaction
Added parser for StringExtra to get the thing name and display it as standard value.

Add strings.xml actions for sensors and devices to choose from, and display either depending on if the thing is a sensor or a device.

Add a button to update gestures, when called it calls on getGestureBoolArray to then be parsed as a bool array in java and append a text string with the fingers, and if they are up or down and display in the TextView for the user.